[#5] refactor: map 컴포넌트 폴더 홈페이지 폴더 내로 이동
